Fixes for VS2019 compile errors. Enough to get it to build and start-up.

+++ kermit/k95/kui/kappwin.cxx
@@ -14,7 +14,7 @@
 #include <process.h>
 #include <sys/stat.h>
 extern "C" {
-    extern int vmode;
+    extern BYTE vmode;
     extern char exedir[];
     extern int  tt_status[];
 };
